C2: promote MapPositionResult to shared/contracts (decouple v1 from V3 feature)
