+++ apps/server/src/shared/contracts/index.ts
@@ -3,5 +3,6 @@ export * from "./envelope.schema";
 export * from "./error.schema";
 export * from "./pagination.schema";
 export * from "./position.schema";
+export * from "./protocol-adapter";
 export * from "./token.schema";
 export * from "./token-amount.schema";
