+++ apps/server/src/shared/contracts/protocol-adapter.ts
@@ -1,0 +1,13 @@
+import type { Position } from "./position.schema";
+
+export interface TokenMetaInput {
+  chainId: number;
+  address: string;
+  symbol: string;
+  decimals: number;
+}
+
+export interface MapPositionResult {
+  position: Position;
+  tokenMetaInputs: TokenMetaInput[];
+}
